Remove 'rugged' and use the 'git' gem.
Openshift deployments are not actually in a Git repo!  Instead the
repo is checked out as a bare repo in the home directory.  Rugged
didn't have the ability to look at the git log without just walking
through all the commits.  The 'git' gem just shells out to the git
executable instead.  Probably more fragile, but at least it works.