Update Candlepin Setup howto to include pgsql help
Prior to this patch, the howto for candlepin left the posgresql
configuration as an exercise left to the reader - however it wasn't very
clear that was the case. There were helpful troubleshooting tips, but
not instructions for the happy path on EL6.

This commit takes the instructions from the developer deployment page
and adds them to this howto, as they are needed to make an EL6 candlepin
deployment functional.

Signed-off-by: Michael Stahnke <stahnma@fedoraproject.org>