Add line numbering option for code excerpts.
Line numbers are rendered with CSS based on anchor tags inserted by
Pygments.

This patch also changes the DOM structure slightly for code excerpts.
In order to get Pygments to insert line number anchors, we must allow it
to wrap the code in a div and pre.  Previously we were instructing
Pygments not to provide any wrapping at all.