Modify build hooks.
Do not force clean builds via marker file (this functionality is now
available in rhc); do not run a bundle install ourselves; and allow
rack-jekyll to build the site instead of building it while the cartridge
is stopped.

The advantage of this system is that rack-jekyll is configured to
display a page alerting users that the site is being rendered instead of
just a sparse Apache error page.  The disadvantage is that the OpenShift
build script will never fail so we don't have a clear idea of whether
the site had any deployment problems or not.  Travis does not currently
fail a build if there are deployment errors, so at least this change
does not leave us in a worse position.