Static Swagger Documentation
Created project folder 'swagger' that uses an empty 'swagger' layout. This way,
the Swagger UI is served statically. Swagger descriptor (swagger.json) is the
input for the static documentation. It has to be pushed into repository when we
want to update the documentation. The swagger.json files for Candlepin will
reside in swagger/candlepin folder. We can keep multiple Swagger files for
different versions of Candlepin  in there and modify links to Swagger UI so that
the link shows correct version of the documentation.

I also deprecated '5.1 Rest API' and created a new Swagger API page with
information about runtime behavior of Swagger integration and links to static
Swagger documentation.