Correct issue with site deployment related to encoding issues.
A new developer joined the team with diacritics in his name.  His name
appeared in the Git log which our site inspects to determine the last
time a page was modified.  Our application was running under an ASCII
locale which meant the UTF-8 characters in the git log caused a runtime
exception.

Rack-jekyll was masking the error because of a bug it had where it would
eternally serve the "please wait" page if rendering failed.  A proposed
fix for that issue is in https://github.com/adaoraul/rack-jekyll/pull/50

This PR should correct the underlying issue with the git log inspection
(i.e. use a UTF-8 locale) and re-enable the rack-jekyll "please wait"
page since the underlying issue has been addressed.