Add Dockerfile and s2i files
The Dockerfile is used to build a base container for us that has Java,
graphviz, and Ruby all of which are underlying dependencies.  The
Dockerfile also copies in PlantUML and the s2i scripts.

The s2i scripts are used by s2i to create a build image and to define
how to run the container that results from the build image.  In our
case, the build image is almost the same as a stock Ruby build image
except we run 'jekyll build' to generate the site.