Remove Nokogiri. Update container image name
We haven't actually used Nokogiri for ages.  A long time ago, we used it
for a custom Jekyll plugin that rendered the JSON definition of our API.
That plugin was removed a long time ago, but this dependency wasn't.

The container image for Centos' Ruby 2.7 container had a name change.